corrected loop condition in ptohex fct get_size

+++ itohex.c
@@ -6,13 +6,13 @@
 /*   By: tkleynts <tkleynts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/10 14:21:02 by tkleynts          #+#    #+#             */
-/*   Updated: 2019/12/10 14:21:10 by tkleynts         ###   ########.fr       */
+/*   Updated: 2020/01/20 15:44:20 by tkleynts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static size_t			get_size(unsigned int num)
+static size_t			get_size(unsigned long long num)
 {
 	size_t	size;
 
@@ -25,12 +25,12 @@ static size_t			get_size(unsigned int num)
 	return (size);
 }
 
-char					*itohex(int num, char *base)
+char					*itohex(unsigned long long num, char *base)
 {
 	char	*string;
 	size_t	size;
 
-	size = get_size((unsigned int)num);
+	size = get_size((unsigned long long)num);
 	string = (char *)malloc(size + 1);
 	string[size] = '\0';
 	while (size)
